# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
}

export interface AuthSession {
  token: string;
  user: AuthUser;
}

export interface InspectionInput {
  /** @minLength 1 */
  payloadText: string;
}

export type ThreatSeverity = typeof ThreatSeverity[keyof typeof ThreatSeverity];


export const ThreatSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface Threat {
  type: string;
  value: string;
  severity: ThreatSeverity;
  rationale: string;
}

export type InspectionResultSeverity = typeof InspectionResultSeverity[keyof typeof InspectionResultSeverity];


export const InspectionResultSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface InspectionResult {
  id: number;
  sanitizedText: string;
  severity: InspectionResultSeverity;
  threats: Threat[];
  rationale: string;
  createdAt: string;
}

export type AuditLogSeverity = typeof AuditLogSeverity[keyof typeof AuditLogSeverity];


export const AuditLogSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface AuditLog {
  id: number;
  preview: string;
  severity: AuditLogSeverity;
  threatCount: number;
  piiCount: number;
  createdAt: string;
}

export type AuditStatsSeverityCounts = {
  LOW: number;
  MEDIUM: number;
  HIGH: number;
};

export type AuditStatsTimelineItem = {
  label: string;
  scans: number;
  risk: number;
};

export interface AuditStats {
  totalScans: number;
  piiIntercepted: number;
  riskScore: number;
  severityCounts: AuditStatsSeverityCounts;
  timeline: AuditStatsTimelineItem[];
}

export type GetAuditLogsParams = {
severity?: GetAuditLogsSeverity;
};

export type GetAuditLogsSeverity = typeof GetAuditLogsSeverity[keyof typeof GetAuditLogsSeverity];


export const GetAuditLogsSeverity = {
  ALL: 'ALL',
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;