# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogSeverity } from './auditLogSeverity';

export interface AuditLog {
  id: number;
  preview: string;
  severity: AuditLogSeverity;
  threatCount: number;
  piiCount: number;
  createdAt: Date;
}