# lib/api-zod/src/generated/types/auditStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditStatsSeverityCounts } from './auditStatsSeverityCounts';
import type { AuditStatsTimelineItem } from './auditStatsTimelineItem';

export interface AuditStats {
  totalScans: number;
  piiIntercepted: number;
  riskScore: number;
  severityCounts: AuditStatsSeverityCounts;
  timeline: AuditStatsTimelineItem[];
}