# lib/api-zod/src/generated/types/inspectionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InspectionResultSeverity } from './inspectionResultSeverity';
import type { Threat } from './threat';

export interface InspectionResult {
  id: number;
  sanitizedText: string;
  severity: InspectionResultSeverity;
  threats: Threat[];
  rationale: string;
  createdAt: Date;
}