# lib/api-zod/src/generated/types/threat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatSeverity } from './threatSeverity';

export interface Threat {
  type: string;
  value: string;
  severity: ThreatSeverity;
  rationale: string;
}